@@ -25,4 +25,4 @@ require (
 	golang.org/x/text v0.7.0 // indirect
 )
 
-go 1.18
+go 1.19
